+++ src/CleanRegex/Internal/Match/Stream/Upstream.php
@@ -21,9 +21,6 @@ interface Upstream
      * @throws EmptyStreamException Supposed to be thrown
      * when the stream is empty and there is no first element,
      * for example when it was filtered out.
-     * @throws StreamRejectedException Supposed to be thrown
-     * when first element of the stream couldn't be resolved,
-     * because of a match reason (unmatched subject or group).
      */
     public function first(): array;
 }